requirements_v2.txt
fastapi==0.115.0
uvicorn==0.27.1
pydantic==2.11.7
requests==2.32.3
pdfplumber==0.11.4
google-generativeai==0.8.2
numpy==1.26.4
scikit-learn==1.5.1
nltk==3.8.1
python-multipart==0.0.9
psutil==5.9.8
PyPDF2==3.0.1 